// packages/schemas/src/database/outputTypeSchemas/BibEntryFindManyArgsSchema.ts
export const BibEntrySelectSchema: z.ZodType<Prisma.BibEntrySelect> = z.object({
  id: z.boolean().optional(),
  BibId: z.boolean().optional(),
  OwnWork: z.boolean().optional(),
  ColleaguesWork: z.boolean().optional(),
  read: z.boolean().optional(),
  htmlCitation: z.boolean().optional(),
  PdfPath: z.boolean().optional(),
  address: z.boolean().optional(),
  annote: z.boolean().optional(),
  author: z.boolean().optional(),
  booktitle: z.boolean().optional(),
  chapter: z.boolean().optional(),
  crossref: z.boolean().optional(),
  doi: z.boolean().optional(),
  edition: z.boolean().optional(),
  editor: z.boolean().optional(),
  email: z.boolean().optional(),
  howpublished: z.boolean().optional(),
  institution: z.boolean().optional(),
  journal: z.boolean().optional(),
  month: z.boolean().optional(),
  note: z.boolean().optional(),
  number: z.boolean().optional(),
  organization: z.boolean().optional(),
  pages: z.boolean().optional(),
  publisher: z.boolean().optional(),
  school: z.boolean().optional(),
  series: z.boolean().optional(),
  title: z.boolean().optional(),
  volume: z.boolean().optional(),
  type: z.boolean().optional(),
  year: z.boolean().optional(),
  numpages: z.boolean().optional(),
  url: z.boolean().optional(),
  issue: z.boolean().optional(),
  issn: z.boolean().optional(),
  abstract: z.boolean().optional(),
  urldate: z.boolean().optional(),
  keywords: z.boolean().optional(),
  copyright: z.boolean().optional(),
  createdAt: z.boolean().optional(),
  lastAccess: z.boolean().optional(),
  Bib: z.union([z.boolean(),z.lazy(() => BibArgsSchema)]).optional(),
  citationGroups: z.union([z.boolean(),z.lazy(() => CitationsGroupFindManyArgsSchema)]).optional(),
  tags: z.union([z.boolean(),z.lazy(() => TagFindManyArgsSchema)]).optional(),
  topics: z.union([z.boolean(),z.lazy(() => TopicFindManyArgsSchema)]).optional(),
  subjects: z.union([z.boolean(),z.lazy(() => SubjectFindManyArgsSchema)]).optional(),
  MdxNotes: z.union([z.boolean(),z.lazy(() => MdxNoteFindManyArgsSchema)]).optional(),
  ipynbNotes: z.union([z.boolean(),z.lazy(() => IpynbFindManyArgsSchema)]).optional(),
  readingList: z.union([z.boolean(),z.lazy(() => ReadingListFindManyArgsSchema)]).optional(),
  _count: z.union([z.boolean(),z.lazy(() => BibEntryCountOutputTypeArgsSchema)]).optional(),
}).strict()

export const BibEntryFindManyArgsSchema: z.ZodType<Prisma.BibEntryFindManyArgs> = z.object({
  select: BibEntrySelectSchema.optional(),
  include: BibEntryIncludeSchema.optional(),
  where: BibEntryWhereInputSchema.optional(),
  orderBy: z.union([ BibEntryOrderByWithRelationInputSchema.array(),BibEntryOrderByWithRelationInputSchema ]).optional(),
  cursor: BibEntryWhereUniqueInputSchema.optional(),
  take: z.number().optional(),
  skip: z.number().optional(),
  distinct: z.union([ BibEntryScalarFieldEnumSchema,BibEntryScalarFieldEnumSchema.array() ]).optional(),
}).strict() ;